// game-client/assets/scripts/app/GeneratedAudioUrls.ts
// Generated by build-wechat-experience.ps1 - do not edit manually.
export const GeneratedAudioUrls = {
  bgm: {
    lobbyAmbient: 'https://assets.flybirdmahjong.fun/game-assets/remote/resources/native/d6/d618c15b-eed1-4ab3-af88-e7182de43ecc.b52cb.mp3',
    tableAmbient: 'https://assets.flybirdmahjong.fun/game-assets/remote/resources/native/b4/b408e93e-ce16-4d41-88ae-6f855203254b.9a968.mp3',
  },
};